PROJQUAY-12493: fix(go): remove change-visibility API handler
The POST /api/v1/repository/{name}/changevisibility endpoint is not
needed for OMR v3.0. Remove the HTTP handler, route registration, and
Service interface method while leaving the underlying service/DAL code
intact. Rename visibility.go → repository.go since the file now only
contains the delete handler.
