+++ internal/api/v1/repository/repository_test.go
@@ -19,7 +19,6 @@ import (
 
 	apiv1 "github.com/quay/quay/internal/api/v1"
 	"github.com/quay/quay/internal/auth"
-	"github.com/quay/quay/internal/dal/daldb"
 	repomodel "github.com/quay/quay/internal/repository"
 	repositorydal "github.com/quay/quay/internal/repository/dal"
 	_ "modernc.org/sqlite"
@@ -295,12 +294,7 @@ func (staticAuthenticator) Authenticate(*http.Request) auth.Result {
 }
 
 type stubService struct {
-	changeErr error
 	deleteErr error
-}
-
-func (s stubService) ChangeVisibility(context.Context, *auth.Principal, repomodel.Ref, repomodel.Visibility) error {
-	return s.changeErr
 }
 
 func (s stubService) Delete(context.Context, *auth.Principal, repomodel.Ref) error {
@@ -328,282 +322,6 @@ func captureLogs(t *testing.T) *bytes.Buffer {
 		slog.SetDefault(previous)
 	})
 	return &buf
-}
-
-func TestChangeVisibilityAdminSuccess(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandler(t, db)
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/public/publicrepo/changevisibility", strings.NewReader(`{"visibility":"public"}`))
-	setBasicAuth(req, "admin")
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusOK {
-		t.Fatalf("status = %d, body = %s", rec.Code, rec.Body.String())
-	}
-	if got := repositoryVisibility(t, db, "public", "publicrepo"); got != "public" {
-		t.Fatalf("visibility = %q, want public", got)
-	}
-}
-
-func TestChangeVisibilityOwnerSuccessWithSlashRepository(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandler(t, db)
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/devtable/repo/withslash/changevisibility", strings.NewReader(`{"visibility":"public"}`))
-	setBasicAuth(req, "devtable")
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusOK {
-		t.Fatalf("status = %d, body = %s", rec.Code, rec.Body.String())
-	}
-	if got := repositoryVisibility(t, db, "devtable", "repo/withslash"); got != "public" {
-		t.Fatalf("visibility = %q, want public", got)
-	}
-}
-
-func TestChangeVisibilityDirectRepositoryAdminSuccess(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandler(t, db)
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/public/publicrepo/changevisibility", strings.NewReader(`{"visibility":"public"}`))
-	setBasicAuth(req, "directadmin")
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusOK {
-		t.Fatalf("status = %d, body = %s", rec.Code, rec.Body.String())
-	}
-	if got := repositoryVisibility(t, db, "public", "publicrepo"); got != "public" {
-		t.Fatalf("visibility = %q, want public", got)
-	}
-}
-
-func TestChangeVisibilityTeamRepositoryAdminSuccess(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandler(t, db)
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/public/publicrepo/changevisibility", strings.NewReader(`{"visibility":"public"}`))
-	setBasicAuth(req, "teamadmin")
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusOK {
-		t.Fatalf("status = %d, body = %s", rec.Code, rec.Body.String())
-	}
-	if got := repositoryVisibility(t, db, "public", "publicrepo"); got != "public" {
-		t.Fatalf("visibility = %q, want public", got)
-	}
-}
-
-func TestChangeVisibilityOrgAdminTeamSuccess(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandler(t, db)
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/public/publicrepo/changevisibility", strings.NewReader(`{"visibility":"public"}`))
-	setBasicAuth(req, "orgadmin")
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusOK {
-		t.Fatalf("status = %d, body = %s", rec.Code, rec.Body.String())
-	}
-	if got := repositoryVisibility(t, db, "public", "publicrepo"); got != "public" {
-		t.Fatalf("visibility = %q, want public", got)
-	}
-}
-
-func TestChangeVisibilityConfiguredSuperuserSuccess(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandlerWithSuperUsers(t, db, []string{"superadmin"}, true)
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/public/publicrepo/changevisibility", strings.NewReader(`{"visibility":"public"}`))
-	setBasicAuth(req, "superadmin")
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusOK {
-		t.Fatalf("status = %d, body = %s", rec.Code, rec.Body.String())
-	}
-	if got := repositoryVisibility(t, db, "public", "publicrepo"); got != "public" {
-		t.Fatalf("visibility = %q, want public", got)
-	}
-}
-
-func TestChangeVisibilityConfiguredSuperuserRequiresFullAccess(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandlerWithSuperUsers(t, db, []string{"superadmin"}, false)
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/public/publicrepo/changevisibility", strings.NewReader(`{"visibility":"public"}`))
-	setBasicAuth(req, "superadmin")
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusForbidden {
-		t.Fatalf("status = %d, want 403", rec.Code)
-	}
-	if got := repositoryVisibility(t, db, "public", "publicrepo"); got != "private" {
-		t.Fatalf("visibility = %q, want unchanged private", got)
-	}
-}
-
-func TestChangeVisibilityRequestBodyTooLarge(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandler(t, db)
-
-	body := `{"visibility":"` + strings.Repeat("x", changeVisibilityRequestMaxBytes) + `"}`
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/public/publicrepo/changevisibility", strings.NewReader(body))
-	setBasicAuth(req, "admin")
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusRequestEntityTooLarge {
-		t.Fatalf("status = %d, want 413", rec.Code)
-	}
-	if got := repositoryVisibility(t, db, "public", "publicrepo"); got != "private" {
-		t.Fatalf("visibility = %q, want unchanged private", got)
-	}
-}
-
-func TestChangeVisibilityUnexpectedErrorLogs(t *testing.T) {
-	sentinelErr := errors.New("database unavailable")
-	logs := captureLogs(t)
-	handler := newStubHandler(t, stubService{changeErr: sentinelErr})
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/public/publicrepo/changevisibility", strings.NewReader(`{"visibility":"public"}`))
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusInternalServerError {
-		t.Fatalf("status = %d, want 500", rec.Code)
-	}
-	logOutput := logs.String()
-	if !strings.Contains(logOutput, "visibility update failed") || !strings.Contains(logOutput, sentinelErr.Error()) {
-		t.Fatalf("log output = %q, want visibility update failure with underlying error", logOutput)
-	}
-}
-
-func TestChangeVisibilityRequiresAuth(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandler(t, db)
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/public/publicrepo/changevisibility", strings.NewReader(`{"visibility":"public"}`))
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusUnauthorized {
-		t.Fatalf("status = %d, want 401", rec.Code)
-	}
-	if got := rec.Header().Get("WWW-Authenticate"); got == "" {
-		t.Fatal("expected WWW-Authenticate header")
-	}
-}
-
-func TestChangeVisibilityForbidden(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandler(t, db)
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/public/publicrepo/changevisibility", strings.NewReader(`{"visibility":"public"}`))
-	setBasicAuth(req, "reader")
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusForbidden {
-		t.Fatalf("status = %d, want 403", rec.Code)
-	}
-}
-
-func TestChangeVisibilityInvalidVisibility(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandler(t, db)
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/public/publicrepo/changevisibility", strings.NewReader(`{"visibility":"internal"}`))
-	setBasicAuth(req, "admin")
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusBadRequest {
-		t.Fatalf("status = %d, want 400", rec.Code)
-	}
-}
-
-func TestSetVisibilityInvalidVisibilityNotFound(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	store := repositorydal.NewStore(db)
-
-	repo, err := daldb.New(db).GetRepositoryAccessByNamespaceName(t.Context(), daldb.GetRepositoryAccessByNamespaceNameParams{
-		Username: "public",
-		Name:     "publicrepo",
-	})
-	if err != nil {
-		t.Fatalf("lookup repository: %v", err)
-	}
-
-	err = store.SetVisibility(t.Context(), repo.ID, repomodel.Visibility("internal"))
-	if !errors.Is(err, repomodel.ErrNotFound) {
-		t.Fatalf("err = %v, want ErrNotFound", err)
-	}
-	if got := repositoryVisibility(t, db, "public", "publicrepo"); got != "private" {
-		t.Fatalf("visibility = %q, want unchanged private", got)
-	}
-}
-
-func TestChangeVisibilityMissingRepository(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandler(t, db)
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/public/missing/changevisibility", strings.NewReader(`{"visibility":"public"}`))
-	setBasicAuth(req, "admin")
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusNotFound {
-		t.Fatalf("status = %d, want 404", rec.Code)
-	}
-}
-
-func TestChangeVisibilityDeletedRepositoryNotFound(t *testing.T) {
-	db := setupAPITestDB(t)
-	defer db.Close()
-	handler := newAPITestHandler(t, db)
-
-	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/api/v1/repository/devtable/deleted/changevisibility", strings.NewReader(`{"visibility":"public"}`))
-	setBasicAuth(req, "admin")
-	rec := httptest.NewRecorder()
-
-	handler.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusNotFound {
-		t.Fatalf("status = %d, want 404", rec.Code)
-	}
 }
 
 func TestDeleteRepositoryAdminSuccess(t *testing.T) {
@@ -732,19 +450,6 @@ func TestDeleteRepositoryMissingRepository(t *testing.T) {
 func setBasicAuth(req *http.Request, username string) {
 	token := base64.StdEncoding.EncodeToString([]byte(username + ":password"))
 	req.Header.Set("Authorization", "Basic "+token)
-}
-
-func repositoryVisibility(t *testing.T, db *sql.DB, namespace, repository string) string {
-	t.Helper()
-	q := daldb.New(db)
-	repo, err := q.GetRepositoryAccessByNamespaceName(t.Context(), daldb.GetRepositoryAccessByNamespaceNameParams{
-		Username: namespace,
-		Name:     repository,
-	})
-	if err != nil {
-		t.Fatalf("lookup repository: %v", err)
-	}
-	return repo.Visibility
 }
 
 func assertRepositoryDeleted(t *testing.T, db *sql.DB, originalName string) {
